Update README.md
Add DSi stuff and some grammar/markdown adjustments

I also mentioned here that the sound/music files must be on the root of the SD card in the DSi/3DS section. I'm not sure if this was an oversight in the readme or the code, but since that is the way it is now.